Extract `RabbitMqVhostExistenceEnsurer` from `EnsureVhostExistsCommand` command.
