+++ README.md
@@ -26,10 +26,10 @@ producers:
 ```
 The value of the `service_alias` should be provided in the `rabbitmq-queue-bundle.default_rabbitmq_producer_name` parameter.
 
-* 'rabbitmq-queue-bundle.rabbitmq-user'
-* 'rabbitmq-queue-bundle.rabbitmq-password'
-* 'rabbitmq-queue-bundle.rabbitmq-vhost' E.g. `/`
-* 'rabbitmq-queue-bundle.rabbitmq-api-base-url' E.g. `http://localhost:15672`. This is used only by commands, that
+* 'rabbitmq-queue-bundle.rabbitmq_user'
+* 'rabbitmq-queue-bundle.rabbitmq_password'
+* 'rabbitmq-queue-bundle.rabbitmq_vhost' E.g. `/`
+* 'rabbitmq-queue-bundle.rabbitmq_api_base_url' E.g. `http://localhost:15672`. This is used only by commands, that
  call the rabbit management api.
  
 ### Important notice: Use dedicated EntityManager for your consumers.
